requirements.txt
azure-storage-blob==12.4.0
